modules: create a separate module for the flow reader
